created the DocsLayout component
