+++ packages/app/src/pages/[user]/index.tsx
@@ -1,12 +1,8 @@
-import { Block, Flex } from '@cube-dev/ui-kit';
 import { getMDXComponent } from 'mdx-bundler/client';
 import { GetServerSideProps, GetServerSidePropsContext, InferGetServerSidePropsType } from 'next';
 import { useRouter } from 'next/router';
 import { useMemo } from 'react';
-import Sidebar from '../../components/docs/Sidebar';
-import Layout from '../../components/shared/Layout';
-import { ScrollSpy } from '../../components/utils/scrollSpy';
-import Container from '../../components/utils/utils';
+import DocsLayout from '../../components/shared/Layouts/DocsLayout';
 import { getConfig } from '../../lib/getConfig';
 import { getMdxContent } from '../../lib/getMDXContent';
 
@@ -17,7 +13,6 @@ export async function getServerSideProps(context: GetServerSidePropsContext) {
   const page = await getMdxContent(username, repo);
   const config = await getConfig(username, repo);
 
-  console.log(username, repo);
   let TOC = JSON.parse(config).navigation;
   return {
     props: {
@@ -46,29 +41,14 @@ function Docs({
   const Component = useMemo(() => getMDXComponent(parsedPage.code), [parsedPage.code]);
 
   return (
-    <Layout>
-      <Block color="#fff">
-        <Container flow="row" display="flex">
-          <Block hide={[false, false, true]}>
-            <Sidebar repoName={repo} repoOwner={username} TOC={TOC} />
-          </Block>
-          <Flex
-            flow="column"
-            gap="2.5rem"
-            className="Outpost-generated"
-            flex="1"
-            margin="0 0 100px 0"
-            padding={['50px 40px', '50px 40px', '20px 0']}
-          >
-            <h1>{parsedPage.frontmatter.title}</h1>
-            <Component />
-          </Flex>
-          <Block hide={[false, true]}>
-            <ScrollSpy />
-          </Block>
-        </Container>
-      </Block>
-    </Layout>
+    <DocsLayout
+      pageTitle={parsedPage.frontmatter.title}
+      TOC={TOC}
+      repoName={username}
+      repoOwner={repo}
+    >
+      <Component />
+    </DocsLayout>
   );
 }
